+++ _unittests/ut_mlmodel/test_piecewise_regressor.py
@@ -7,7 +7,7 @@ import numpy
 from numpy.random import random
 import pandas
 from sklearn.linear_model import LinearRegression
-from pyquickhelper.pycode import ExtTestCase
+from pyquickhelper.pycode import ExtTestCase, ignore_warnings
 from mlinsights.mlmodel import test_sklearn_pickle, test_sklearn_clone, test_sklearn_grid_search_cv
 from mlinsights.mlmodel.piecewise_estimator import PiecewiseRegressor
 
@@ -37,6 +37,7 @@ class TestPiecewiseRegressor(ExtTestCase):
         self.assertNotEqual(pred2.min(), pred2.max())
         self.assertGreater(clq.n_estimators_, 1)
 
+    @ignore_warnings(UserWarning)
     def test_piecewise_regressor_no_intercept_bins(self):
         X = numpy.array([[0.1, 0.2], [0.2, 0.3], [0.2, 0.35], [0.2, 0.36]])
         Y = numpy.array([1., 1.1, 1.15, 1.2])